[service] rename substrateApi to js-api-wrappers
