+++ lib/service/js-api-wrappers/api.dart
@@ -4,12 +4,12 @@ import 'dart:convert';
 import 'package:encointer_wallet/config/node.dart';
 import 'package:encointer_wallet/service/ipfsApi/httpApi.dart';
 import 'package:encointer_wallet/service/subscan.dart';
-import 'package:encointer_wallet/service/substrateApi/apiAccount.dart';
-import 'package:encointer_wallet/service/substrateApi/apiAssets.dart';
-import 'package:encointer_wallet/service/substrateApi/chainApi.dart';
-import 'package:encointer_wallet/service/substrateApi/codecApi.dart';
-import 'package:encointer_wallet/service/substrateApi/encointer/apiEncointer.dart';
-import 'package:encointer_wallet/service/substrateApi/types/genExternalLinksParams.dart';
+import 'package:encointer_wallet/service/js-api-wrappers/apiAccount.dart';
+import 'package:encointer_wallet/service/js-api-wrappers/apiAssets.dart';
+import 'package:encointer_wallet/service/js-api-wrappers/chainApi.dart';
+import 'package:encointer_wallet/service/js-api-wrappers/codecApi.dart';
+import 'package:encointer_wallet/service/js-api-wrappers/encointer/apiEncointer.dart';
+import 'package:encointer_wallet/service/js-api-wrappers/types/genExternalLinksParams.dart';
 import 'package:encointer_wallet/store/app.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
