+++ lib/service/js-api-wrappers/types/genExternalLinksParams.dart
